pyproject.toml
[project]
name = "tileatlas-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "networkx[default]>=3.6.1",
    "orjson>=3.11.7",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
]
